+++ ReleaseNotes.md
@@ -2,6 +2,11 @@
 
 These are the NuGet package releases. See also [npm Release Notes](ReleaseNotesNpm.md).
 
+## 3.8.0
+
+* Set explicit Content-Type headers in fastify plugin routes.
+* Revert "Send explicit return values" from 3.7.0
+
 ## 3.7.0
 
 * Add `4xx` and `5xx` response schemas to fastify plugin.